Move EVM to Status folder
